Windows fixes

+++ src/CommunitiesList.cpp
@@ -104,7 +104,8 @@ CommunitiesList::setTagsForRoom(const QString &room_id, const std::vector<std::s
                 }
                 // insert or remove the room from the tag as appropriate
                 std::string current_tag =
-                  it->first.right(it->first.size() - strlen("tag:")).toStdString();
+                  it->first.right(static_cast<int>(it->first.size() - strlen("tag:")))
+                    .toStdString();
                 if (std::find(tags.begin(), tags.end(), current_tag) != tags.end()) {
                         // the room has this tag
                         it->second->addRoom(room_id);
@@ -246,7 +247,7 @@ CommunitiesList::fetchCommunityAvatar(const QString &id, const QString &avatarUr
 
                   cache::saveImage(opts.mxc_url, res);
 
-                  auto data = QByteArray(res.data(), res.size());
+                  auto data = QByteArray(res.data(), (int)res.size());
 
                   QPixmap pix;
                   pix.loadFromData(data);
